added the needed functionality

+++ src/main/java/dk/sb/userrating/RatingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.Collection;
 
 public interface RatingRepository extends MongoRepository<Rating, String> {
-    Collection<Rating> findAllByCar(int id);
+    Collection<Rating> findAllByCar(Long id);
     Collection<Rating> findAllByUser(User user);
 }
